+++ src/utils/Str.ts
@@ -7,3 +7,14 @@ export const trim = (s, c) => {
     '',
   );
 };
+
+export const addSlashes = (string) => {
+	return string.replace(/\\/g, '\\\\').
+			replace(/\u0008/g, '\\b').
+			replace(/\t/g, '\\t').
+			replace(/\n/g, '\\n').
+			replace(/\f/g, '\\f').
+			replace(/\r/g, '\\r').
+			replace(/'/g, '\\\'').
+			replace(/"/g, '\\"');
+}
