+++ tests/Unit/cli/cli.test.ts
@@ -1,32 +1,9 @@
-import { spawn } from 'child_process';
 import { DiagramBuilder } from '../../../src/server/DiagramBuilder';
 import CreateJSON from '../../../src/server/nodes/CreateJSON'
 import Inspect from '../../../src/server/nodes/Inspect'
 import { nonCircularJsonStringify } from '../../../src/utils/nonCircularJsonStringify';
-
-/** SPAWN A CHILD PROCESS AND CAPTURE STDOUT AND STDERR */
-async function spawnChild(command, arguments_, options) {
-	const child = spawn(command, arguments_, options);
-
-	let data = "";
-	for await (const chunk of child.stdout) {
-			console.log('stdout chunk: '+chunk);
-			data += chunk;
-	}
-	let error = "";
-	for await (const chunk of child.stderr) {
-			console.error('stderr chunk: '+chunk);
-			error += chunk;
-	}
-	const exitCode = await new Promise( (resolve, reject) => {
-			child.on('close', resolve);
-	});
-
-	if( exitCode) {
-			throw new Error( `subprocess error exit ${exitCode}, ${error}`);
-	}
-	return data;
-}
+import { addSlashes } from '../../../src/utils/Str'
+import { spawnChild } from '../../../src/utils/process_'
 
 async function cli($command, ...$args) {
 	return await spawnChild(
@@ -40,17 +17,6 @@ async function cli($command, ...$args) {
 			cwd: __dirname
 		}
 	)
-}
-
-function addslashes(string) {
-	return string.replace(/\\/g, '\\\\').
-			replace(/\u0008/g, '\\b').
-			replace(/\t/g, '\\t').
-			replace(/\n/g, '\\n').
-			replace(/\f/g, '\\f').
-			replace(/\r/g, '\\r').
-			replace(/'/g, '\\\'').
-			replace(/"/g, '\\"');
 }
 
 it('can boot via cli', async () => {
@@ -86,7 +52,7 @@ it.skip('can run via cli', async () => {
 	}
 
 	let json = nonCircularJsonStringify(diagram)
-	json = addslashes(json)
+	json = addSlashes(json)
 
   await cli('run "' + json + '"').then(success, fail)
 });
